move HighlightedText component to components/text
