+++ web-ui/src/components/text/highlight-text/highlight-text.less
